Before oauth

+++ src/main/java/io/github/platovd/authserver/jwt/JWTResponse.java
@@ -4,7 +4,7 @@ public record JWTResponse(
         String accessToken,
         String refreshToken,
         String tokenType,
-        Integer expiresIn,
-        Integer refreshExpiresIn
+        Long expiresIn,
+        Long refreshExpiresIn
 ) {
 }
